+++ src/main/java/dev/booky/stackdeobf/mixin/CrashReportMixin.java
@@ -1,6 +1,7 @@
 package dev.booky.stackdeobf.mixin;
 // Created by booky10 in StackDeobfuscator (18:35 20.03.23)
 
+import dev.booky.stackdeobf.mappings.RemappedThrowable;
 import dev.booky.stackdeobf.mappings.RemappingUtil;
 import net.minecraft.CrashReport;
 import org.spongepowered.asm.mixin.Final;
@@ -10,6 +11,7 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(CrashReport.class)
 public class CrashReportMixin {
@@ -22,5 +24,21 @@ public class CrashReportMixin {
     )
     public void postInit(String title, Throwable throwable, CallbackInfo ci) {
         this.exception = RemappingUtil.remapThrowable(throwable);
+    }
+
+    @Inject(
+            method = "getException",
+            at = @At("HEAD"),
+            cancellable = true
+    )
+    public void preExceptionGet(CallbackInfoReturnable<Throwable> cir) {
+        // redirect calls to getException to the original, unmapped Throwable
+        //
+        // this method is called in the ReportedException, which
+        // caused the "RemappedThrowable" name to show up in the logger
+
+        if (this.exception instanceof RemappedThrowable remapped) {
+            cir.setReturnValue(remapped.getOriginal());
+        }
     }
 }